Extract printClass and printTypeParameters
